\name{findrange.alt}
\alias{findrange.alt}

\title{
    Extract a range for depiction of an alt object in a relationship plot
}
\description{
    Extract a range for depiction of an alt object in a relationship plot.
	}
\usage{findrange.alt(x)
	}

\arguments{
    \item{x}{
        An \code{"alt"} object as generated by \code{"alt.make"}.
	}
}

\details{
	All $data elements and any $goal element are scanned to identify the bounds of data entry.
	If element $parallel_par exists the P1 values will also be scanned for possible expansion of the bounds.
}

\value{
	A data frame will be returned with columns denoting yrange and xrange.	
}

\references{

  Wayne Nelson, (1990) "Accelerated Testing", Wiley-Interscience, New York
  
  William Q. Meeker and Luis A. Escobar, (1998) "Statistical Methods for Reliability Data", Wiley-Interscience, New York
  
  Pascual F., Meeker W., Escobar L. (2006) Accelerated Life Test Models and Data Analysis. In: Pham H. (eds) Springer Handbook of Engineering Statistics. Springer Handbooks. Springer, London. https://doi.org/10.1007/978-1-84628-288-1_22
}
